change id from long to int
# Conflicts:
#	src/main/java/com/learnx/demo/dao/UserDAO.java
#	src/main/java/com/learnx/demo/model/User.java

+++ src/main/java/com/learnx/demo/service/AppUserService.java
@@ -12,5 +12,5 @@ public interface AppUserService {
 
     AppUserDto getUserById(int userId);
 
-    List<AppUserDto> listInstructorsByInstituteId(long instituteId);
+    List<AppUserDto> listInstructorsByInstituteId(int instituteId);
 }
